requirements.txt
aiosmtplib==2.0.1
alembic==1.10.2
anyio==3.6.2
-e git+https://github.com/d-exponent/expense-tracker.git@bd2c71a44110bb5d28245ef3c937437c542a2433#egg=app
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
orjson==3.8.7
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.1.1
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.6
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
tomli==2.0.1
tqdm==4.65.0
twilio==7.17.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
